requirements.txt
Django==5.0.6
djangorestframework>=3.11.0
djoser>=2.0.0
django-environ==0.11.2
psycopg2-binary==2.9.9
